+++ android/quest/src/main/java/org/smartregister/fhircore/quest/configuration/view/PatientRegisterRowViewConfiguration.kt
@@ -39,6 +39,7 @@ data class Filter(
   val key: String,
   val displayableProperty: String = key,
   val valuePrefix: String? = null,
+  val valuePostfix: String? = null,
   val label: String? = null,
   val valueType: Enumerations.DataType,
   val valueCoding: Code? = null,