+++ android/quest/src/main/java/org/smartregister/fhircore/quest/data/patient/PatientRepository.kt
@@ -23,23 +23,40 @@ import com.google.android.fhir.search.StringFilterModifier
 import com.google.android.fhir.search.search
 import javax.inject.Inject
 import kotlinx.coroutines.withContext
+import org.hl7.fhir.r4.context.SimpleWorkerContext
+import org.hl7.fhir.r4.model.Base
 import org.hl7.fhir.r4.model.CodeableConcept
+import org.hl7.fhir.r4.model.Coding
+import org.hl7.fhir.r4.model.Condition
 import org.hl7.fhir.r4.model.Encounter
+import org.hl7.fhir.r4.model.Enumerations
+import org.hl7.fhir.r4.model.Observation
 import org.hl7.fhir.r4.model.Patient
 import org.hl7.fhir.r4.model.Questionnaire
 import org.hl7.fhir.r4.model.QuestionnaireResponse
+import org.hl7.fhir.r4.model.Resource
+import org.hl7.fhir.r4.model.ResourceType
+import org.hl7.fhir.r4.model.StringType
+import org.hl7.fhir.r4.utils.FHIRPathEngine
 import org.smartregister.fhircore.engine.configuration.ConfigurationRegistry
 import org.smartregister.fhircore.engine.configuration.view.SearchFilter
 import org.smartregister.fhircore.engine.data.domain.util.PaginationUtil
 import org.smartregister.fhircore.engine.data.domain.util.RegisterRepository
 import org.smartregister.fhircore.engine.ui.questionnaire.QuestionnaireConfig
 import org.smartregister.fhircore.engine.util.DispatcherProvider
+import org.smartregister.fhircore.engine.util.extension.asDdMmmYyyy
 import org.smartregister.fhircore.engine.util.extension.countActivePatients
-import org.smartregister.fhircore.quest.configuration.parser.DetailConfigParser
+import org.smartregister.fhircore.engine.util.extension.getEncounterId
+import org.smartregister.fhircore.engine.util.extension.referenceValue
+import org.smartregister.fhircore.engine.util.extension.valueToString
+import org.smartregister.fhircore.quest.configuration.view.Filter
 import org.smartregister.fhircore.quest.configuration.view.PatientDetailsViewConfiguration
+import org.smartregister.fhircore.quest.configuration.view.isSimilar
+import org.smartregister.fhircore.quest.data.patient.model.AdditionalData
 import org.smartregister.fhircore.quest.data.patient.model.PatientItem
 import org.smartregister.fhircore.quest.data.patient.model.QuestResultItem
 import org.smartregister.fhircore.quest.ui.patient.register.PatientItemMapper
+import org.smartregister.fhircore.quest.util.getSearchResults
 import org.smartregister.fhircore.quest.util.loadAdditionalData
 import timber.log.Timber
 
@@ -51,6 +68,8 @@ constructor(
   private val dispatcherProvider: DispatcherProvider,
   val configurationRegistry: ConfigurationRegistry
 ) : RegisterRepository<Patient, PatientItem> {
+
+  private val fhirPathEngine = FHIRPathEngine(SimpleWorkerContext())
 
   override suspend fun loadData(
     query: String,
@@ -93,24 +112,127 @@ constructor(
   suspend fun fetchTestResults(
     patientId: String,
     forms: List<QuestionnaireConfig>,
-    patientDetailsViewConfiguration: PatientDetailsViewConfiguration,
-    parser: DetailConfigParser?
+    config: PatientDetailsViewConfiguration
   ): List<QuestResultItem> {
     return withContext(dispatcherProvider.io()) {
       val testResults = mutableListOf<QuestResultItem>()
 
-      parser?.let { p ->
-        val questionnaireResponses =
-          searchQuestionnaireResponses(patientId, forms).sortedByDescending { it.authored }
-        questionnaireResponses.forEach {
-          val questionnaire = getQuestionnaire(it)
-          testResults.add(p.getResultItem(questionnaire, it, patientDetailsViewConfiguration))
-        }
+      val questionnaireResponses =
+        searchQuestionnaireResponses(patientId, forms).sortedByDescending { it.authored }
+      questionnaireResponses.forEach {
+        val questionnaire = getQuestionnaire(it)
+        testResults.add(getResultItem(questionnaire, it, config))
       }
-        ?: run { Timber.w("Getting null parser") }
-
       testResults
     }
+  }
+
+  suspend fun getResultItem(
+    questionnaire: Questionnaire,
+    questionnaireResponse: QuestionnaireResponse,
+    patientDetailsViewConfiguration: PatientDetailsViewConfiguration
+  ): QuestResultItem {
+
+    when {
+      patientDetailsViewConfiguration.dynamicRows.isEmpty() -> {
+        val data =
+          listOf(
+            listOf(
+              AdditionalData(value = fetchResultItemLabel(questionnaire)),
+              AdditionalData(value = " (${questionnaireResponse.authored?.asDdMmmYyyy() ?: ""})")
+            )
+          )
+
+        return QuestResultItem(Pair(questionnaireResponse, questionnaire), data)
+      }
+      else -> {
+
+        val encounter = loadEncounter(questionnaireResponse.getEncounterId())
+        val data: MutableList<List<AdditionalData>> = mutableListOf()
+
+        patientDetailsViewConfiguration.dynamicRows.forEach { filtersList ->
+          val additionalDataList = mutableListOf<AdditionalData>()
+
+          filtersList.forEach { filter ->
+            val value =
+              when (filter.resourceType) {
+                Enumerations.ResourceType.CONDITION ->
+                  getCondition(encounter, filter)
+                    .find { doesSatisfyFilter(it, filter) == true }
+                    ?.getPathValue(filter.displayableProperty)
+                Enumerations.ResourceType.OBSERVATION ->
+                  getObservation(encounter, filter)
+                    .find { doesSatisfyFilter(it, filter) == true }
+                    ?.getPathValue(filter.displayableProperty)
+                else -> null
+              }
+
+            additionalDataList.add(
+              AdditionalData(
+                label = filter.label,
+                value = value.valueToString(),
+                valuePrefix = filter.valuePrefix,
+                valuePostfix = filter.valuePostfix,
+                properties = filter.properties
+              )
+            )
+          }
+          data.add(additionalDataList)
+        }
+
+        return QuestResultItem(Pair(questionnaireResponse, questionnaire), data)
+      }
+    }
+  }
+
+  fun Base.getPathValue(path: String) = fhirPathEngine.evaluate(this, path).firstOrNull()
+
+  fun doesSatisfyFilter(resource: Resource, filter: Filter): Boolean? {
+    if (filter.valueCoding == null && filter.valueString == null)
+      throw IllegalStateException("Filter must have either of one valueCoding or valueString")
+
+    // get property mentioned as filter and match value
+    // e.g. category: CodeableConcept in Condition
+    return resource
+      .getNamedProperty(filter.key)
+      .values
+      .firstOrNull()
+      ?.let {
+        when (it) {
+          // match relevant type and value
+          is CodeableConcept -> filter.valueCoding!!.isSimilar(it)
+          is Coding -> filter.valueCoding!!.isSimilar(it)
+          is StringType -> it.value == filter.valueString
+          else -> false
+        }
+      }
+      .also {
+        if (it == null)
+          Timber.i("${resource.resourceType}, ${filter.key}: could not resolve key value filter")
+      }
+  }
+
+  suspend fun getCondition(encounter: Encounter, filter: Filter?) =
+    getSearchResults<Condition>(encounter.referenceValue(), Condition.ENCOUNTER, filter, fhirEngine)
+      .sortedByDescending {
+        if (it.hasOnsetDateTimeType()) it.onsetDateTimeType.value else it.recordedDate
+      }
+      .sortedByDescending { it.logicalId }
+
+  suspend fun getObservation(encounter: Encounter, filter: Filter?) =
+    getSearchResults<Observation>(
+      encounter.referenceValue(),
+      Observation.ENCOUNTER,
+      filter,
+      fhirEngine
+    )
+      .sortedByDescending {
+        if (it.hasEffectiveDateTimeType()) it.effectiveDateTimeType.value else it.meta.lastUpdated
+      }
+      .sortedByDescending { it.logicalId }
+
+  fun fetchResultItemLabel(questionnaire: Questionnaire): String {
+    return questionnaire.name ?: questionnaire.title ?: questionnaire.logicalId
   }
 
   suspend fun getQuestionnaire(questionnaireResponse: QuestionnaireResponse): Questionnaire {